requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sounddevice==0.4.6
numpy==1.26.2
boto3==1.29.3
scipy==1.11.3
python-dotenv==1.0.0
gspread==5.12.0
google-auth==2.23.4
requests==2.31.0
pandas==2.1.1
openpyxl==3.1.2 